+++ CHANGES.rst
@@ -1,6 +1,7 @@
 0.1.1
 -----
 * Bug fix: IStaticResource fails to look up self.request if nested 2-deep
+* Bug fix: name collisions with version_helper.py
 
 0.1.0
 -----